requirements.txt
kornia
opencv-python
redis
omegaconf
hydra-core
imageio
git+https://github.com/facebookresearch/pytorch3d.git@stable
gdown 
transformers
scikit-image
submitit
addict
imageio
matplotlib 
plotly 
opencv-python
pytorch_lightning==1.8.6
einops
trimesh
wandb
chumpy
numpy==1.23.1 ## required by chumpy
setuptools==59.5.0
smplx
git+https://github.com/hassony2/manopth.git
projectaria-tools'[all]'